feat: jwt 검증 로직 추가 및 패키지 위치 변경
